+++ chameleon_openid/tests.py
@@ -3,6 +3,7 @@ from django.core.urlresolvers import reverse
 from django.test import TestCase
 from django.utils.importlib import import_module
 import mock
+import json
 
 class SessionTestCase(TestCase):
     def setUp(self):
@@ -14,6 +15,23 @@ class SessionTestCase(TestCase):
 
 class OpenIDViewsTests(SessionTestCase):
 
+    def setUp(self):
+        super(OpenIDViewsTests, self).setUp()
+        self.openid_session = {
+            'status': 'testing',
+            'url': 'http://example.com/openid.php?uri=testing',
+            'sreg': {
+                'nickname': 'jdoe1',
+                'email': 'jdoe1@example.com',
+            },
+            'ax': {
+                'projects': [],
+                'full_name': 'John Doe',
+            }
+        }
+        self.session['openid'] = self.openid_session
+        self.session.save()
+
     @mock.patch('tas.forms.get_country_choices')
     @mock.patch('tas.forms.get_department_choices')
     @mock.patch('tas.forms.get_institution_choices')
@@ -24,21 +42,6 @@ class OpenIDViewsTests(SessionTestCase):
         mock_get_department_choices.return_value = ((127, 'Texas Advanced Computing Center'),)
         mock_get_country_choices.return_value = ((230, 'United States'),)
 
-        result = {}
-        result['status'] = 'testing'
-        result['url'] = 'http://example.com/openid.php?uri=testing'
-        result['sreg'] = {
-            'nickname': 'jdoe1',
-            'email': 'jdoe1@example.com',
-        }
-        result['ax'] = {
-            'projects': [],
-            'full_name': 'John Doe',
-        }
-
-        self.session['openid'] = result
-        self.session.save()
-
         response = self.client.get(reverse('chameleon_openid:openid_register'))
         self.assertEqual(response.status_code, 200)
         self.assertContains(response, \
@@ -46,3 +49,24 @@ class OpenIDViewsTests(SessionTestCase):
         self.assertContains(response, \
             '<input class="form-control" id="id_email" name="email" placeholder="Email"'
             ' required="required" title="" type="email" value="jdoe1@example.com" />')
+
+    @mock.patch('pytas.http.TASClient.save_user')
+    @mock.patch('tas.forms.get_country_choices')
+    @mock.patch('tas.forms.get_department_choices')
+    @mock.patch('tas.forms.get_institution_choices')
+    def test_openid_register_submit(self, mock_get_institution_choices,
+                                  mock_get_department_choices, mock_get_country_choices,
+                                  mock_tas_save_user):
+
+        mock_get_institution_choices.return_value = ((1, 'The University of Texas at Austin'),)
+        mock_get_department_choices.return_value = ((127, 'Texas Advanced Computing Center'),)
+        mock_get_country_choices.return_value = ((230, 'United States'),)
+        mock_tas_save_user.return_value = json.loads(open(
+            'chameleon_openid/test_fixtures/user.json').read())
+        form_data = json.loads(open(
+            'chameleon_openid/test_fixtures/registration_form.json').read())
+
+        response = self.client.post(reverse('chameleon_openid:openid_register'), form_data)
+
+        self.assertTrue(mock_tas_save_user.called)
+